Fix uhtml+chrome keyed node bug
Basically it's textarea blur event is fired when a sch node is removed,
is the problem here. When we rename a sch key, editSelected will
render(store) which will remove and add a new sch to the same position.
That node removal also fires a blur event, and we also render(store)
in onblur callback, at this point the removing node is not connected,
thus uhtml throws exception of that reason.

The solution we chose here is not letting the node be removed by not
losing its sch reference, so uhtml knows it's the same keyed node and
reuse it!