Redo All aria-selected behaviours
This should be good enough generally. There are two issues not being
addressed at this point: 1) Meta+click AND THEN shift+arrow is still
the same as shift+arrow behaviour while MacOS is inclusively select
as shift+arrow moves. 2) We brought back firefox slowness at repainting
bottom to top for sake of simplicity for now, we will fix this later