Fix cloneUp/Down due to unserializable store
Store was having _allSch property that contains a list of functions,
that broke result of cloneUp/Down because those type(sch) constructors
were lost.

So we always keep store or regular sch object serializable from now on.
However, do not over use Sch.clone except for when we need new reference
of object.