Default prefix key as _tag (all _box is now _tag)
Tag concept is for semantic meaning of sch object. We keep those types
as primitives, no customization. We hope to reuse this concept on ui gen
with reusing code base.

We tend to not have clear convention around underscore naming (_tag).
Currently we think they are for non-schema, non-persistent, ui-only
sort of thing.

For HTMLElement object, we always use underscore to avoid prop name
conflict since we don't control the object.