Aria spec says aria-expanded is at treeitem nodes
[TreeView](https://www.w3.org/TR/wai-aria-practices/#TreeView) quote:

Each element with role treeitem that serves as a parent node
has aria-expanded set to false when the node is in a closed state
and set to true when the node is in an open state.

End nodes do not have the aria-expanded attribute because,
if they were to have it, they would be incorrectly described
to assistive technologies as parent nodes.

---

While I think [role='group'] owned by treeitem makes more sense,
we do whatever aria spec says