Fmodels count and fix addDefault (enumerable related)
We should test thoese enumerable: false related soon, it gives us
anxiety around sch mutation (e.g. clone and spread like in this commit)

Also spot bug when move fmodels across file and one of item has
path equal (or substring) to dstPath, and causes items not to pop
out because it implies we intend to move into subtree. So when we
know it's a move across file, we pass empty string (ops("")) as a
dstPath, it should _always_ pop out because "" does not start with
any string except node.id == "" which is root item (we can't move that
anyway)