Fix keyed node, it strikes again
The problem here is indexed container's prepend child can't edit
its type (no parent error), I guess this is because path that ties
to index is problematic, what was at 0 is now at 1, and the new one is
at 0, it seems to reuse node when we use "path" as reference.

We now use html.for(assigns.parent, assigns.sch.$anchor) again,
IIRC, there was also a problem for this, but as we test a bunch of
actions today, there's no show problem.