Fix all tagged_diff weirdness stuff
Maybe "not all" but it's more clear now than ever. We were scare of
using non-serializable object inside Sch.walk meta, now we go for
lpath (list of path) that store sch instances along the way, it's easier
to refer to without re-finding it again. Under non-enumberable "_meta",
it should be safe.