Fix tagged_diff mutaul exclusion diff
Our assumption was that there was no Diff.NEW_KEY and Diff.MOVED at the
same time, that's wrong. This scenario can happen. Moving a subfmodel
having the same key as a child of new parent resulting in the moving
item changing its key so all children of new parent are unique.