Make pkgs/model as a subpath package
This is different than the standlone. Unlike a standlone package
that include core, it shares core modules with the core.

In short, standalone is single purpose only. Once you need more than
one package, use the suppath version.

We may get rid of standlone in future if most of the time the
sup-package does the job.

Note that other pkgs/* will be done in next commit