Fix undefined _meta.index inside structsheet checker
We need to know that index but we don't want to query store every time
so without store query, the _meta is not attached.

Maybe we need to implement full linked list traveral from a particular
sch node to get basic property like index (via parent), next|prev sch
and so on.