Remove pinned-dst mechanism. It was dumb.
It walks store to put a pin in dst and then walks againt to get pin
out. Probably because I thought to not depend on `$a` at some point.
Now `$a` or id is required everywhere. Pin is no longer needed.

The reason we had pinned-dst was because of path based (indexed) problem
next adjacent index was (i-1) when moving previous adjacant into it.
pop(dst-i) -> put(dst-i-1) so pin was to remember dst (not using stale
path id)